rename to serve-git
